feat(metrics): add layout and integrate API data
